Chance license to ZlibWithAcknowledgement
s.a. `https://spdx.org/licenses/zlib-acknowledgement.html'
I felt like the standard ZLib license was written waaaay to much in favor of using my work in closed source products, and really was only hindering the open-source community as a whole (which I've really grown to be a huge fan of in recent times), by requiring any changes made to marked as such (which I still find to be quite the fair way of doing things), but then going on to say: If your product isn't going to be open-source, you're not even required to admit to using my work for it (i.e. having that part be optional).

Anyways: This new license simply made the acknowledgement-part in closed-source products legally mandatory (though I would like to add that personally I'm OK with open-source products not adding explicit documenation about my contributions, as I find the derived source itself (which was already required to include references to my work) to be ~documentation~ (to use the term found in the new licnese header) enough)