(mostly) implement a new magic-generator system
This new system is designed much more modular and is essentially meant to replace the abomination that /kos/misc/magicgenerator/generate_header.dee has been turned into over the time since its inception. Note though that this new system has yet to be _fully_ completed, nor would it be able to parse _all_ of the existing magic source files in its current state. However, once complete, this new system will fix all of the flaws of current magic (the worst of which is (was): It being impossible to read, or modify its convoluted source code, as well as wrong decisions in how ertain header portions are generated).
The decision was made after I've noticed that the (old) magic system caused wchar functions to get bound incorrectly for PE-mode, and I'm no longer able to understand the voodoo it does to cause that is causing this error.

It's not usually my style to commit partially completed work, however after working on this on and off for the last week already (with the end only slowly coming into sight, and over 215K bytes of new source code already written (where the old magic generator is 301K bytes)), it makes sense to create this commit of about ~70% of the work done.