Use unicode data for <wctype.h>
Also change the backing tables for <ctype.h> to only be applicable to ASCII (rather than LATIN-1). Functions from <ctype.h> are meant to be used for single-byte characters, and given that KOS uses UTF-8 for this purpose, that means that only the ASCII sub-set exists in single-byte form. Byts 80h-FFh aren't latin-1 characters, but incomplete UTF-8 sequence bytes, which is now being reflected by <ctype.h> functions indicate none of the properties for such bytes.