Implement the new regex compiler
I really think this is going to work; I did have to adjust the design a little bit to fix some flaws, and I've yet to implement the compiler for character-sets, but everything else should already compile correctly (!UNTESTED!).

So with that in mind, I guess the next step will be to finish implementation of the missing compiler features, as well as implement `regexec(3)`!